requirements.txt
PyQt5~=5.15.10
requests~=2.32.3
pillow~=10.3.0